refactor: SpillManager into a separate file (#15407)
* refactor: SpillManager into a separate file

* chore
